move channel metadata export function to its own file; autoformat; safertyping
