+++ sigexport/merge.py
@@ -1,3 +1,4 @@
+import datetime
 import re
 import shutil
 from pathlib import Path
@@ -45,10 +46,10 @@ def merge_chat(new: list[models.Message], path_old: Path) -> list[models.Message
     # get rid of duplicates
     msg_dict = {m.comp(): m for m in old_msgs + new}
     merged = list(msg_dict.values())
-    
-    def get_date(val: any):
+
+    def get_date(val: models.Message) -> datetime.datetime:
         return val.date
-    
+
     merged.sort(key=get_date)
 
     return merged
